+++ xolotl/interface.h
@@ -84,8 +84,8 @@ public:
 	 * @param zs, zm The start and width in the Z direction on the local MPI process
 	 * @param Mz The total width in the Z direction
 	 */
-	void getLocalCoordinates(int &xs, int &xm, int &Mx, int &ys, int &ym,
-			int &My, int &zs, int &zm, int &Mz);
+	void getLocalCoordinates(long int &xs, long int &xm, long int &Mx, long int &ys, long int &ym,
+			long int &My, long int &zs, long int &zm, long int &Mz);
 
 	/**
 	 * Set the location of one GB grid point.
