+++ xolotl/tests/modifiedreaction/OneReSolutionHandlerTester.cpp
@@ -106,10 +106,10 @@ BOOST_AUTO_TEST_CASE(checkReSolution) {
 
 	// Initialize the indices and values to set in the Jacobian
 	int nXenon = reSolutionHandler.getNumberOfReSoluting();
-	int indices[10 * nXenon];
+	long int indices[10 * nXenon];
 	double val[10 * nXenon];
 	// Get the pointer on them for the compute re-solution method
-	int *indicesPointer = &indices[0];
+	long int *indicesPointer = &indices[0];
 	double *valPointer = &val[0];
 
 	// Compute the partial derivatives for the re-solution at the grid point 8
@@ -229,10 +229,10 @@ BOOST_AUTO_TEST_CASE(checkMinimumSize) {
 
 	// Initialize the indices and values to set in the Jacobian
 	int nXenon = reSolutionHandler.getNumberOfReSoluting();
-	int indices[10 * nXenon];
+	long int indices[10 * nXenon];
 	double val[10 * nXenon];
 	// Get the pointer on them for the compute re-solution method
-	int *indicesPointer = &indices[0];
+	long int *indicesPointer = &indices[0];
 	double *valPointer = &val[0];
 
 	// Compute the partial derivatives for the re-solution at the grid point 8
@@ -343,10 +343,10 @@ BOOST_AUTO_TEST_CASE(checkDifferentFit) {
 
 	// Initialize the indices and values to set in the Jacobian
 	int nXenon = reSolutionHandler.getNumberOfReSoluting();
-	int indices[10 * nXenon];
+	long int indices[10 * nXenon];
 	double val[10 * nXenon];
 	// Get the pointer on them for the compute re-solution method
-	int *indicesPointer = &indices[0];
+	long int *indicesPointer = &indices[0];
 	double *valPointer = &val[0];
 
 	// Compute the partial derivatives for the re-solution at the grid point 8
