+++ xolotl/xolotlCore/advection/DummyAdvectionHandler.h
@@ -81,7 +81,7 @@ public:
 	 * \see IAdvectionHandler.h
 	 */
 	void computePartialsForAdvection(const IReactionNetwork& network,
-			double *val, int *indices, const NDPoint<3>& pos, double hxLeft,
+			double *val, long int *indices, const NDPoint<3>& pos, double hxLeft,
 			double hxRight, int ix, double hy = 0.0, int iy = 0,
 			double hz = 0.0, int iz = 0) const override {
 		// Doesn't do anything
