+++ xolotl/xolotlCore/advection/IAdvectionHandler.h
@@ -114,7 +114,7 @@ public:
 	 * @param iz The position on the z grid
 	 */
 	virtual void computePartialsForAdvection(const IReactionNetwork& network,
-			double *val, int *indices, const NDPoint<3>& pos, double hxLeft,
+			double *val, long int *indices, const NDPoint<3>& pos, double hxLeft,
 			double hxRight, int ix, double hy = 0.0, int iy = 0,
 			double hz = 0.0, int iz = 0) const = 0;
