+++ xolotl/xolotlCore/advection/YGBAdvectionHandler.cpp
@@ -135,7 +135,7 @@ void YGBAdvectionHandler::computeAdvection(const IReactionNetwork& network,
 }
 
 void YGBAdvectionHandler::computePartialsForAdvection(
-		const IReactionNetwork& network, double *val, int *indices,
+		const IReactionNetwork& network, double *val, long int *indices,
 		const NDPoint<3>& pos, double hxLeft, double hxRight, int ix, double hy,
 		int iy, double hz, int iz) const {
