+++ xolotl/xolotlCore/diffusion/Diffusion2DHandler.cpp
@@ -116,7 +116,7 @@ void Diffusion2DHandler::computeDiffusion(const IReactionNetwork& network,
 }
 
 void Diffusion2DHandler::computePartialsForDiffusion(
-		const IReactionNetwork& network, double *val, int *indices,
+		const IReactionNetwork& network, double *val, long int *indices,
 		double hxLeft, double hxRight, int ix, double sy, int iy, double,
 		int) const {
