+++ xolotl/xolotlCore/modifiedreaction/heterogeneousnucleation/IHeterogeneousNucleationHandler.h
@@ -83,7 +83,7 @@ public:
 	 * @return true if nucleation is happening
 	 */
 	virtual bool computePartialsForHeterogeneousNucleation(const IReactionNetwork& network,
-			double *val, int *indices, int xi, int xs, int yj = 0,
+			double *val, long int *indices, int xi, int xs, int yj = 0,
 			int zk = 0) = 0;
 
 };
