+++ xolotl/xolotlCore/modifiedreaction/resolution/DummyReSolutionHandler.h
@@ -89,7 +89,7 @@ public:
 	 * \see IReSolutionHandler.h
 	 */
 	int computePartialsForReSolution(const IReactionNetwork& network,
-			double *val, int *indices, int xi, int xs, int yj = 0, int zk = 0) {
+			double *val, long int *indices, int xi, int xs, int yj = 0, int zk = 0) {
 		return 0;
 	}
