+++ xolotl/xolotlCore/modifiedreaction/resolution/OneReSolutionHandler.h
@@ -359,7 +359,7 @@ public:
 	 * \see IReSolutionHandler.h
 	 */
 	int computePartialsForReSolution(const IReactionNetwork &network,
-			double *val, int *indices, int xi, int xs, int yj = 0, int zk = 0) {
+			double *val, long int *indices, int xi, int xs, int yj = 0, int zk = 0) {
 		// Get the single xenon which connects to every cluster
 		auto singleXenon = network.get(Species::Xe, 1);
 		int xenonId = singleXenon->getId() - 1;
