+++ xolotl/xolotlCore/modifiedreaction/trapmutation/TrapMutationHandler.h
@@ -201,7 +201,7 @@ public:
 	 * \see ITrapMutationHandler.h
 	 */
 	int computePartialsForTrapMutation(const IReactionNetwork& network,
-			double *val, int *indices, int xi, int yj = 0, int zk = 0);
+			double *val, long int *indices, int xi, int yj = 0, int zk = 0);
 
 	/**
 	 * Get the total number of clusters in the network that can undergo trap mutation.
