+++ xolotl/xolotlCore/reactants/alloyclusters/AlloyClusterReactionNetwork.cpp
@@ -538,7 +538,7 @@ void AlloyClusterReactionNetwork::computeAllFluxes(double *updatedConcOffset,
 }
 
 void AlloyClusterReactionNetwork::computeAllPartials(
-		const std::vector<size_t>& startingIdx, const std::vector<int>& indices,
+		const std::vector<size_t>& startingIdx, const std::vector<long int>& indices,
 		std::vector<double>& vals, int i) const {
 
 	// Initial declarations
