+++ xolotl/xolotlCore/temperature/ITemperatureHandler.h
@@ -108,7 +108,7 @@ public:
 	 * @param sz The space parameter, depending on the grid step size in the z direction
 	 * @param iz The position on the z grid
 	 */
-	virtual void computePartialsForTemperature(double *val, int *indices,
+	virtual void computePartialsForTemperature(double *val, long int *indices,
 			double hxLeft, double hxRight, int xi, double sy = 0.0, int iy = 0,
 			double sz = 0.0, int iz = 0) = 0;
