+++ xolotl/xolotlCore/temperature/TemperatureGradientHandler.h
@@ -145,7 +145,7 @@ public:
 	 *
 	 * \see ITemperatureHandler.h
 	 */
-	virtual void computePartialsForTemperature(double *val, int *indices,
+	virtual void computePartialsForTemperature(double *val, long int *indices,
 			double hxLeft, double hxRight, int xi, double sy = 0.0, int iy = 0,
 			double sz = 0.0, int iz = 0) {
 		// Set the cluster index, the PetscSolver will use it to compute
